requirements.txt
# Core ML/DL
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

# Quantum Computing
pennylane==0.38.0
qiskit==1.2.0

# API Backend
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Frontend/UI
streamlit==1.38.0
plotly==5.24.0

# Image Processing
albumentations==1.4.18
opencv-python==4.10.0.84
Pillow>=10.0.0

# Data & Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Experiment Tracking
wandb==0.17.8

# Async Processing
celery==5.4.0
redis==5.0.8

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.32

# Validation
pydantic==2.9.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Progress bars
tqdm>=4.66.0

# Production
gunicorn==22.0.0
